requirements.txt
torch>=2.6.0
torchvision
transformers>=4.50.0
bitsandbytes>=0.45.0
accelerate>=1.0.0
peft>=0.14.0
qwen-vl-utils
datasets
pillow
